requirements.txt
absl-py==2.3.0
accelerate==1.8.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
backrefs==5.9
bcrypt==3.2.0
beautifulsoup4==4.13.4
beniget==0.4.1
black==21.4b2
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.5.2
camelot-py==1.0.9
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.0.3
click-plugins==1.1.1
cloudpickle==3.1.1
cohere==5.15.0
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
command-not-found==0.3
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
cssselect==1.3.0
cssutils==2.11.1
cupshelpers==1.0
cycler==0.11.0
Cython==3.1.2
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.16
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docstring_parser==0.16
duplicity==0.8.21
effdet==0.4.1
einops==0.8.1
email_validator==2.2.0
emoji==2.14.1
entrypoints==0.4
et_xmlfile==2.0.0
etelemetry==0.3.1
evdev==1.9.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastavro==1.11.1
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.29.1
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.7.0
fs==2.4.12
fsspec==2025.5.1
ftfy==6.3.1
ftz==0.1
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.5.2
GDAL==3.4.1
ghp-import==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==1.34.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==1.0.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.49.0rc1
gyp==0.1
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
identify==2.6.12
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib-metadata==4.6.4
importlib_resources==6.5.2
instructor==1.9.0
iopath==0.1.9
ipykernel==6.30.1
ipython==8.37.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.1
language-selector==0.1
lark==1.2.2
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
looseversion==1.3.0
louis==3.20.0
lxml==4.8.0
lz4==4.4.4
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.5.1
mypy_extensions==1.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
overrides==7.7.0
OWSLib==0.25.0
packaging==25.0
paddleocr==3.0.2
paddlepaddle==3.0.0
paddlex==3.0.2
paginate==0.5.7
pandas==2.3.0
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
patsy==1.0.1
pbr==5.8.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pi_heif==0.22.0
pickleshare==0.7.5
pikepdf==9.9.0
pillow==11.3.0
platformdirs==4.3.8
ply==3.11
polars==1.34.0
polars-runtime-32==1.34.0
portalocker==3.2.0
pprintpp==0.4.0
pre_commit==4.2.0
premailer==3.10.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
protoc-gen-openapiv2==0.0.1
prov==2.1.1
psutil==7.0.0
psycopg2==2.9.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py==1.10.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycairo==1.20.1
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pycups==2.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pygccxml==2.2.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymdown-extensions==10.16
PyMuPDF==1.26.3
pymupdf4llm==0.0.27
PyNaCl==1.5.0
pynput==1.8.1
pynvim==0.4.2
PyOpenGL==3.1.5
pypandoc==1.15
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.1
pyproj==3.3.0
PyQt-Qwt==1.2.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-telegram-bot==22.5
python-xlib==0.33
pythran==0.10.0
pyttsx3==2.99
pytz==2022.1
pyxdg==0.27
pyxnat==1.6.3
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.0.1
QScintilla==2.11.6
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
SecretStorage==3.3.1
Send2Trash==1.8.3
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
srt==3.5.3
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
stringzilla==3.12.5
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
terminator==2.1.1
testpath==0.5.0
threadpoolctl==3.6.0
thrift==0.16.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.52.4
triton==3.3.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.10.0
ultralytics==8.3.206
ultralytics-thop==2.0.17
unattended-upgrades==0.1
unicodedata2==14.0.0
unstructured==0.18.1
unstructured-client==0.37.2
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
unstructured_paddleocr==2.10.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
usb-creator==0.3.7
uvicorn==0.35.0
virtualenv==20.31.2
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==2.0.0
wrapt==1.17.2
xdg==5
xkit==0.0.0
xlrd==2.0.2
xlsxwriter==3.2.5
yacs==0.1.8
yarl==1.20.1
zipp==1.0.0
zstandard==0.23.0
